app/requirements.txt
google-adk
fastmcp
psutil
litellm